Move the new `span` package to an `internal` sub-package
