+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/frescosupport/FrescoBasedReactTextInlineImageSpan.java
@@ -27,7 +27,7 @@ import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.modules.fresco.ReactNetworkImageRequest;
 import com.facebook.react.uimanager.PixelUtil;
 import com.facebook.react.views.image.ImageResizeMode;
-import com.facebook.react.views.text.span.TextInlineImageSpan;
+import com.facebook.react.views.text.internal.span.TextInlineImageSpan;
 
 /**
  * FrescoBasedTextInlineImageSpan is a span for Images that are inside <Text/>. It computes its size
