+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/internal/span/ReactBackgroundColorSpan.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react.views.text.span;
+package com.facebook.react.views.text.internal.span;
 
 import android.text.style.BackgroundColorSpan;
