+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/internal/span/ReactSpan.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react.views.text.span;
+package com.facebook.react.views.text.internal.span;
 
 /*
  * Enables us to distinguish between spans that were added by React Native and spans that were
