+++ packages/react-native/ReactAndroid/src/test/java/com/facebook/react/views/text/CustomLineHeightSpanTest.kt
@@ -8,7 +8,7 @@
 package com.facebook.react.views.text
 
 import android.graphics.Paint
-import com.facebook.react.views.text.span.CustomLineHeightSpan
+import com.facebook.react.views.text.internal.span.CustomLineHeightSpan
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
